e2e/pyproject.toml
[project]
name = "chattbc-e2e"
version = "0.1.0"
description = "chattbc.gg E2E tests"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8,<9",
    "pytest-playwright>=0.6,<1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "S", "T20", "SIM", "RUF"]
ignore = ["S105", "S106"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
